requirements.txt
# Top-level dev dependencies
google-auth
gspread
